util: split up find missing translation code
One file is the wrapper for `index.ts` aka `npm run util`
and the other file is the implementation for that code.
