+++ util/find_missing_translations.ts
@@ -3,16 +3,10 @@ import path from 'path';
 import readline from 'readline';
 import { fileURLToPath } from 'url';
 
-import { Namespace, SubParser } from 'argparse';
-import inquirer from 'inquirer';
-
-import { isLang, Lang, languages } from '../resources/languages';
-import { UnreachableCode } from '../resources/not_reached';
+import { isLang, Lang } from '../resources/languages';
 
 import { walkDirSync } from './file_utils';
 import { findMissing } from './find_missing_timeline_translations';
-
-import { ActionChoiceType } from '.';
 
 // Directory names to ignore when looking for JavaScript files.
 const ignoreDirs = [
@@ -152,7 +146,7 @@ const parseJavascriptFile = (file: string, inputLocales: Lang[]) => {
   });
 };
 
-const findMissingTranslations = async (filter: string, locale: Lang): Promise<void> => {
+export const findMissingTranslations = async (filter: string, locale: Lang): Promise<void> => {
   const files = findAllJavascriptFiles(filter);
   for (const file of files) {
     await findMissing(
@@ -171,74 +165,4 @@ const findMissingTranslations = async (filter: string, locale: Lang): Promise<vo
     );
     parseJavascriptFile(file, [locale]);
   }
-};
-
-type FindMissingTranslationsNamespaceInterface = {
-  'filter': string | null;
-  'locale': string | null;
-};
-
-class FindMissingTranslationsNamespace extends Namespace
-  implements FindMissingTranslationsNamespaceInterface {
-  'filter': string | null;
-  'locale': string | null;
-}
-
-type FindMissingTranslationsInquirerType = {
-  [name in keyof FindMissingTranslationsNamespaceInterface]:
-    FindMissingTranslationsNamespaceInterface[name];
-};
-
-const findMissingTranslationsFunc = (args: Namespace): Promise<void> => {
-  if (!(args instanceof FindMissingTranslationsNamespace))
-    throw new UnreachableCode();
-  const questions = [
-    {
-      type: 'fuzzypath',
-      name: 'filter',
-      message: 'Input a valid trigger JavaScript filename: ',
-      rootPath: 'ui',
-      suggestOnly: true,
-      default: args.filter ?? '',
-      when: () => typeof args.filter !== 'string',
-    },
-    {
-      type: 'list',
-      name: 'locale',
-      message: 'Select a locale: ',
-      choices: languages,
-      default: args.locale,
-      when: () => typeof args.locale !== 'string',
-    },
-  ] as const;
-  return inquirer.prompt<FindMissingTranslationsInquirerType>(questions)
-    .then((answers) => {
-      const filter = answers.filter ?? args.filter;
-      const locale = answers.locale ?? args.locale;
-      if (typeof filter === 'string' && typeof locale === 'string' && isLang(locale))
-        return findMissingTranslations(filter, locale);
-    });
-};
-
-export const registerFindMissingTranslations = (
-  actionChoices: ActionChoiceType,
-  subparsers: SubParser,
-): void => {
-  actionChoices.findTranslations = {
-    name: 'Find missing translations',
-    callback: findMissingTranslationsFunc,
-    namespace: FindMissingTranslationsNamespace,
-  };
-  const findParser = subparsers.addParser('findTranslations', {
-    description: actionChoices.findTranslations.name,
-  });
-
-  findParser.addArgument(['-l', '--locale'], {
-    help: 'The locale to find missing translations for, e.g. de',
-  });
-  findParser.addArgument(['-f', '--filter'], {
-    nargs: '?',
-    type: 'string',
-    help: 'Limits the results to only match specific files/path',
-  });
 };
